Clear the workspace's clippy warnings and rustfmt drift
All ten were pre-existing and every fix is behaviour-preserving:

- astar: drop a u32 -> u32 cast; `unsigned_abs` already returns u32
- growth: `sort_by(|a, b| b.1.cmp(&a.1))` -> `sort_by_key` + `Reverse`.
  Both are stable and descending, so the permutation is identical and
  worldgen output cannot drift
- combat: `chain` takes IntoIterator, so the `.into_iter()` was redundant
- inventory: derive the instance id from `enumerate` instead of a manual
  counter; same sequence from the same reserved block
- orchestrator: `label` is already `&str`, so `&label` was auto-deref'd
- dungeon tests: `<= GRID - 1` -> `< GRID`, and name the door cell-pair
  tuple so the collect target reads

Also fold in the one rustfmt drift, so `cargo fmt --check` and
`cargo clippy --workspace --all-targets` are both clean from a fresh
checkout.

+++ shared/src/pathfinding/astar.rs
@@ -107,9 +107,7 @@ pub fn find_path(
     let h = |x: i32, z: i32, fk: u16| -> u32 {
         let real_f = key_to_floor(fk) as i32;
         let goal_f = goal_floor as i32;
-        (x - gx).unsigned_abs()
-            + (z - gz).unsigned_abs()
-            + (real_f - goal_f).unsigned_abs() as u32 * 2
+        (x - gx).unsigned_abs() + (z - gz).unsigned_abs() + (real_f - goal_f).unsigned_abs() * 2
     };
 
     let start_h = h(sx, sz, start_fk);
